[QntSyn] change representation of tyvars to new TyVar type
Previously, named and unnamed (generated) type variables were handled as
two separate constructors on Type: TVar and TUnif. This system led to
the two being treated differently in stages such as type inference,
which wasn't an ideal system.

This is changed by collapsing these two constructors into a single
constructor, 'TVar TyVar'. 'TyVar' here is a new type with two
constructors: 'TvName' and 'TvUnif', for named and unnamed tyvars
respectively. This should allow handling of type variables in a more
representation-agnostic way in the type checker.